Add files via upload
This is my Personal Portfolio Website created with HTML, CSS, and JavaScript, designed to highlight my skills, education, and projects in a modern and professional way. The site is fully responsive, ensuring a smooth experience across desktops, tablets, and mobile devices. It includes interactive sections such as Home, About Me, Skills & Education, Projects with direct live links, and a Contact Form with validation. Deployed on Netlify, the portfolio provides an accessible and visually appealing showcase of my work, while serving as a central hub for my online presence.